combine descriptor, peripheral, service, characteristic.js into blecentral.js, fix some bugs
