--- src/js/characteristic.js
@@ -1,127 +1,0 @@
-var events = require('events');
-var util = require('util');
-
-var characteristics = require('./characteristics.json');
-
-function Characteristic(bleCentral, peripheralId, serviceUuid, 
-                        uuid, properties) {
-  this._bleCentral = bleCentral;
-  this._peripheralId = peripheralId;
-  this._serviceUuid = serviceUuid;
-
-  this.uuid = uuid;
-  this.name = null;
-  this.type = null;
-  this.properties = properties;
-  this.descriptors = null;
-
-  var characteristic = characteristics[uuid];
-  if (characteristic) {
-    this.name = characteristic.name;
-    this.type = characteristic.type;
-  }
-}
-
-util.inherits(Characteristic, events.EventEmitter);
-
-Characteristic.prototype.toString = function() {
-  return JSON.stringify({
-    uuid: this.uuid,
-         name: this.name,
-         type: this.type,
-         properties: this.properties
-  });
-};
-
-Characteristic.prototype.read = function(callback) {
-  if (callback) {
-    this.once('read', function(data) {
-      callback(null, read);
-    });
-  }
-
-  this._bleCentral.read(
-      this._peripheralId,
-      this._serviceUuid,
-      this.uuid
-      );
-};
-
-Characteristic.prototype.write = function(data, withoutResponse, callback) {
-  if (process.title !== 'browser') { //
-    if (!(data instanceof Buffer) {
-      throw new Error('data must be a Buffer');
-    }
-  }
-
-  if (callback) {
-    this.once('write', function() {
-      callback(null);
-    });
-  }
-
-  this._bleCentral.write(
-      this._peripheralId,
-      this._serviceUuid,
-      this.uuid,
-      data,
-      withoutResponse
-      );
-};
-
-Characteristic.prototype.broadcast = function(broadcast, callback) {
-  if (callback) {
-    this.once('broadcast', function() {
-      callback(null);
-    });
-  }
-
-  this._bleCentral.broadcast(
-      this._peripheralId,
-      this._serviceUuid,
-      this.uuid,
-      broadcast
-      );
-};
-
-Characteristic.prototype.notify = function(notify, callback) {
-  if (callback) {
-    this.once('notify', function() {
-      callback(null);
-    });
-  }
-
-  this._bleCentral.notify(
-      this._peripheralId,
-      this._serviceUuid,
-      this.uuid,
-      notify
-      );
-};
-
-
-Characteristic.prototype.subscribe = function(callback) {
-  this.notify(true, callback);
-};
-
-Characteristic.prototype.unsubscribe = function(callback) {
-  this.notify(false, callback);
-};
-
-Characteristic.discoverDescriptors = function(callback) {
-  if (callback) {
-    this.once('dscriptorsDiscover', function(descriptors) {
-      callback(null, descriptors);
-    });
-  }
-
-  this._bleCentral.discoverDescriptors(
-      this._peripheralId,
-      this._serviceUuid,
-      this.uuid
-      );
-};
-
-module.exports = Characteristic;
-}
-
