--- src/js/service.js
@@ -1,63 +1,0 @@
-var events = require('events');
-var util = require('util');
-
-var services = require('./services.json');
-
-function Service(bleCentral, peripheralId, uuid) {
-  this._bleCentral = bleCentral;
-  this._peripheralId = peripheralId;
-  
-  this.uuid = uuid;
-  this.name = null;
-  this.type = null;
-  this.includedServiceUuids = null;
-  this.characteristics = null;
-
-  var service = services[uuid];
-  if(service) {
-    this.name = service.name;
-    this.type = service.type;
-  }
-}
-
-util.inherits(Service, events.EventEmitter);
-
-Service.prototype.toString = function() {
-  return JSON.stringify({
-    uuid: this.uuid,
-         name: this.name,
-         type: this.type,
-         includedServiceUuids: this.includedServiceUuids
-  });
-};
-
-Service.prototype.discoverIncludedServices = function(serviceUuids, callback) {
-  if (callback) {
-    this.once('includedServicesDiscover', function(includedServiceUuids) {
-      callback(null, includedServiceUuids);
-    });
-  }
-
-  this._bleCentral.discoverIncludedServices(
-      this._peripheralId,
-      this.uuid,
-      serviceUuids
-  );
-};
-
-Service.prototype.discoverCharacteristics = 
-function(characteristicUuids, callback) {
-  if (callback) {
-    this.once('characteristicsDiscover', function(characteristics) {
-      callback(null, characteristics);
-    });
-  }
-
-  this._bleCentral.discoverCharacteristics(
-      this._peripheralId,
-      this.uuid,
-      characteristicUuids
-      );
-};
-
-module.exports = Service;
